+++ ui/packages/consul-ui/app/utils/search/exact.js
@@ -1,0 +1,12 @@
+import PredicateSearch from './predicate';
+
+export default class ExactSearch extends PredicateSearch {
+  predicate(s) {
+    s = s.toLowerCase();
+    return (item = '') =>
+      item
+        .toString()
+        .toLowerCase()
+        .indexOf(s) !== -1;
+  }
+}
